requirements.txt
# Core AI
ollama>=0.3.0
chromadb>=0.4.0
langchain>=0.1.0
langchain-community>=0.0.10

# Memory & Learning
networkx>=3.0
numpy>=1.24.0
scikit-learn>=1.3.0

# System Control
psutil>=5.9.0
watchdog>=3.0.0
schedule>=1.2.0
gitpython>=3.1.0

# Safety & Monitoring
python-dotenv>=1.0.0
pyyaml>=6.0
structlog>=24.0.0
rich>=13.0.0

# Web Interface (optional)
fastapi>=0.100.0
uvicorn>=0.20.0

# Database
sqlite3  # Built into Python
sqlalchemy>=2.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Utilities
aiofiles>=23.0.0
httpx>=0.24.0
python-json-logger>=2.0.0